+++ src/main/java/com/ward/ward_server/controller/AuthController.java
@@ -4,11 +4,8 @@ import com.ward.ward_server.model.LoginRequest;
 import com.ward.ward_server.model.LoginResponse;
 import com.ward.ward_server.security.JwtIssuer;
 import com.ward.ward_server.security.UserPrincipal;
+import com.ward.ward_server.service.AuthService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,24 +14,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class AuthController {
-    private final JwtIssuer jwtIssuer;
-    private final AuthenticationManager authenticationManager;
-    
+    private final AuthService authService;
+
     @PostMapping("/auth/login")
     public LoginResponse login(@RequestBody @Validated LoginRequest request){
-        var authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword())
-        );
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        var principal = (UserPrincipal) authentication.getPrincipal();
-
-        var roles = principal.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .toList();
-
-        var token = jwtIssuer.issue(principal.getUserId(), principal.getEmail(), roles);
-        return LoginResponse.builder()
-                .accessToken(token)
-                .build();
+        return authService.attemtLogion(request.getEmail(), request.getPassword());
     }
 }
