+++ src/main/java/com/ward/ward_server/controller/HelloController.java
@@ -20,4 +20,9 @@ public class HelloController {
         return "If you see this, then you're logged in as user " + principal.getEmail()
                 + " User ID: " + principal.getUserId();
     }
+
+    @GetMapping("/admin")
+    public String admin(@AuthenticationPrincipal UserPrincipal principal) {
+        return "If you see this, then you are an Admin. User ID: " + principal.getUserId();
+    }
 }
