+++ src/main/java/com/ward/ward_server/service/AuthService.java
@@ -1,0 +1,34 @@
+package com.ward.ward_server.service;
+
+import com.ward.ward_server.model.LoginResponse;
+import com.ward.ward_server.security.JwtIssuer;
+import com.ward.ward_server.security.UserPrincipal;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AuthService {
+    private final JwtIssuer jwtIssuer;
+    private final AuthenticationManager authenticationManager;
+    public LoginResponse attemtLogion(String email, String password) {
+        var authentication = authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(email, password)
+        );
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        var principal = (UserPrincipal) authentication.getPrincipal();
+
+        var roles = principal.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .toList();
+
+        var token = jwtIssuer.issue(principal.getUserId(), principal.getEmail(), roles);
+        return LoginResponse.builder()
+                .accessToken(token)
+                .build();
+    }
+}
